Use `em` based margins
Keep text spacing consistent when we scale content up or down at
different viewports. Remove the need for extra media queries.

This brings the govspeak component inline with the styles on frontend
and whitehall that have less tight paragraph and list spacing, and are
more legible for it.

Add spacing between list items, as per older front-end typography.